requirements.txt

Django>=4.2
pillow
pytesseract
thefuzz
